requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
autogen-agentchat==0.6.1
autogen-core==0.6.1
autogen-ext==0.6.1
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
jsonref==1.1.0
openai==1.84.0
opentelemetry-api==1.34.0
pillow==11.2.1
protobuf==5.29.5
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
zipp==3.22.0
pytest==7.4.3
mcp>=1.0.0
fastmcp>=0.5.0
python-dotenv>=1.0.0
